refactor(smart-mic): inline i18n as constants and flatten module layout
