+++ src/features/extensions/smart-mic/smart-mic-qr-hero.tsx
