+++ src/smartmic/components/rec-area.helpers.ts
@@ -1,3 +1,4 @@
+// Helpers specific to rec-area.tsx. Promote to ../helpers/ if a second consumer appears.
 export const formatElapsed = (sec: number): string => {
     const mm = Math.floor(sec / 60);
     const ss = sec % 60;
